fix: address telemetry review findings
